Fix duplicate reach issue, and Jupyter notebook run issue
Update are to sets.py and getAllSets.py

To fix the duplicate issue, I overhauled the "remove_duplicate_sets" function, and added a few new short functions that it calls.

To fix the Jupyter notebook issue (which was actually flagged in a different issue; I'll try to remember to loop in on GitHub) I used some of these pages: https://stackoverflow.com/questions/15411967/how-can-i-check-if-code-is-executed-in-the-ipython-notebook

via Nikki ofc!